refactor(ui): extract model capabilities and deduplicate model list
- Move isFastSupported, isEffortSupported, isXhighEffortAllowed, and
  isMaxEffortAllowed into dedicated modelCapabilities.ts module
- Add explicit extraUsage field to MODELS instead of sniffing labels
- Have command palette import MODELS from ModelSelector instead of
  maintaining a duplicate list
- Move capability tests to modelCapabilities.test.ts
